training/requirements.txt
# Training dependencies for dataset generation and centroid computation
numpy>=1.26
scikit-learn>=1.3
scikit-image>=0.22
Pillow>=10.0
joblib>=1.3
tqdm>=4.66
matplotlib>=3.8

# Optional for Jupyter notebooks
jupyter>=1.0
ipykernel>=6.29
